requirements.txt
appdirs==1.4.4
# apturl==0.5.2
astroid==2.4.2
autopep8==1.5.4
bcrypt==3.1.7
blinker==1.4
# Brlapi==0.7.0
certifi==2019.9.11
chardet==3.0.4
# chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
# command-not-found==0.3
cryptography==2.8
# cupshelpers==1.0
Cython==0.29.21
# dbus-python==1.2.16
# defer==1.0.6
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
distro==1.4.0
# distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==2.1.5
django-braces==1.12.0
django-cors-headers==2.1.0
django-heroku==0.3.1
django-import-export==1.0.0
django-jet==1.0.7
django-jet2==1.0.10
django-materializecss-form==1.1.17
django-webpack-loader==0.5.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
# duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==20.0.4
httplib2==0.14.0
idna==2.7
importlib-metadata==1.7.0
iso8601==0.1.12
isort==5.4.2
jdcal==1.4.1
keyring==18.0.1
# language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
# louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupPy==1.14
MarkupSafe==1.1.0
mccabe==0.6.1
mlpack==3.3.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.1
oauthlib==3.1.0
odfpy==1.4.0
olefile==0.46
openpyxl==3.0.0
pandas==1.1.2
paramiko==2.6.0
pbr==5.5.0
pexpect==4.6.0
Pillow==7.0.0
promise==2.1
protobuf==3.6.1
psycopg2==2.8.4
# pycairo==1.16.2
pycodestyle==2.6.0
# pycups==1.9.73
# PyGObject==3.36.0
PyJWT==1.5.3
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyPDF2==1.26.0
pyRFC3339==1.1
# python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.5
# python-debian===0.1.36ubuntu1
pytz==2018.3
pyxdg==0.26
PyYAML==5.1.2
reportlab==3.5.34
requests==2.19.1
requests-unixsocket==0.2.0
Rx==1.6.1
SecretStorage==2.3.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12
stevedore==3.2.2
# systemd-python==234
tablib==0.14.0
toml==0.10.1
typing==3.6.4
# ubuntu-advantage-tools==20.3
# ubuntu-drivers-common==0.0.0
# ufw==0.36
# unattended-upgrades==0.1
urllib3==1.23
# usb-creator==0.3.7
virtualenv==20.0.34
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
whitenoise==5.2.0
wrapt==1.12.1
# xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==1.0.0